telstra-id-data-synth/requirements.txt
google-cloud-pubsub
scipy
simplejson